test: analytics gate, docs parity, README doctor rewrite, test consolidation
Documents the six reliability/update PostHog events (taxonomy + bounded
volumes), adds the Reliability-incidents insight to the dashboard spec
(validated, provisioner tests green), updates six internal docs for the
reliability release, rewrites the stale README doctor section, and merges
the two remoteMachineModel test files into one.
