+++ docs/features/onboarding-and-settings/desktop-auto-update.md
@@ -60,3 +60,53 @@ watchdog, while the native Squirrel handoff has a separate five-minute watchdog
 so loopback transfer and staging are not mistaken for a stalled quit. Handoff
 timeouts retain the pending-install marker because Squirrel may still complete;
 an explicit updater error clears it.
+
+## Truthful version surfaces
+
+Every version surface reads from one shared snapshot so they can never disagree
+about what is running versus what is staged. `AutoUpdateSnapshot` carries both
+`currentVersion` (the running build) and `latestKnownVersion` (the newest
+version `electron-updater` has observed from its configured feed), plus the
+staged `version`, `parked`, and `autoApplyPending` fields below.
+`useAutoUpdateSnapshot` (`renderer/components/app/useAutoUpdateSnapshot.ts`)
+does the initial `updateGetState()` read and subscribes to `onUpdateEvent`; the
+top-bar pill (`AutoUpdateControl`), the app-shell banner (`AutoUpdateBanner`),
+and the Settings About panel (`AboutSection`) all consume it. About shows the
+running version as "Installed" and `latestKnownVersion` as "Latest", but swaps
+"Installed" to the staged version when a download is `ready` or `parked`, so the
+user sees the version they will get after the next restart rather than a stale
+"you're up to date".
+
+## Transactional install and the parked banner
+
+`quitAndInstall()` is transactional. Before flipping the snapshot to
+`installing` it re-runs `updater.checkForUpdates({ allowReady: true })` to
+confirm the staged installer is still the latest, then persists
+`pendingInstallUpdate` and calls `updater.quitAndInstall(false, true)`. A
+consent that aborts before the native updater can take over does not silently
+vanish: the snapshot records `parked: { reason, at }` where `reason` is a typed
+`AutoUpdateInstallAbortReason` — `refresh_failed`, `install_preflight_failed`,
+`prepare_failed`, `prepare_timeout`, or `handoff_failed`. The app-shell
+`AutoUpdateBanner` renders a parked state as "Update to vX didn't finish —
+Restart to retry" (a parked state wins over a plain `ready` state), and a plain
+`ready` state as "Running vCurrent · vNext is ready", each with a **Restart
+now** action wired to `updateQuitAndInstall()`. Banner dismissal is keyed on a
+stable signature so it reappears when a newer version stages or a fresh abort
+occurs, but stays hidden for an unchanged state.
+
+## Idle auto-apply
+
+When auto-apply is enabled (packaged builds with auto-check on, unless
+`ADE_DISABLE_AUTO_UPDATE_APPLY=1`), a `ready` update is applied on its own once
+the machine is quiet. The service polls the runtime's
+`RuntimeActivitySummary` — `idle` is true only when there are no active agent
+turns and no active work sessions. After the runtime has been continuously idle
+for the idle grace period, the snapshot gets an `autoApplyPending: { deadlineAt }`
+countdown; the `AutoUpdateBanner` renders a "Updating to vX in Ns" toast that
+ticks down each second. Reaching the deadline while still `ready` and idle calls
+the same `quitAndInstall()` path and emits `ade_update_auto_applied`. Any
+renewed activity clears the pending countdown, and an explicit user **Cancel**
+(`updateCancelAutoApply`) sets `autoApplySuppressedUntil` so another countdown
+is not started until that epoch passes. `installing` remains a sticky status
+throughout: the service ignores `update-not-available` / `checking-for-update` /
+`error` while a quitAndInstall is in flight.
